Removed redundant ResourceDefinitions null check
The singleton `resourceDefinitions` is initialized at type construction time and so will never be null bey the time `getInstance` is called. This Change removes the redundant null check. On glancing at the original code it looked like a potential race condition with multiple possible  instances being created.

Note that this pattern is repeated elsewhere, but I'm testing the water with this PR to see if you folks are open to community help before I put any more effort in.